VESC (dev) now reading from VESC and sending to phone.

+++ TripAdvisor.Phone/app/src/main/java/com/skelstar/tripadvisorphone/TripData.kt
@@ -1,3 +1,3 @@
 package com.skelstar.tripadvisorphone
 
-data class TripData(val volts: Float, val amphours: Int)
+data class TripData(val volts: Float, val amphours: Float)
